fix: moved methods to extract keywords into classes
