+++ src/main/java/org/rri/ijTextmate/Helpers/LanguageInformationExtractor/InformationExtractor.java
@@ -1,0 +1,7 @@
+package org.rri.ijTextmate.Helpers.LanguageInformationExtractor;
+
+public interface InformationExtractor {
+    String getExtension();
+
+    ExtractedLanguageInformation getExtractedLanguageInformation();
+}
